+++ examples/prometheus/doc.go
@@ -1,0 +1,5 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+// Package main provides a code sample of the Prometheus exporter.
+package main